chore(maintain): update file names and minor changes
