+++ package.json
@@ -29,6 +29,8 @@
     "@babel/polyfill": "^7.4.4",
     "@sendgrid/mail": "^6.4.0",
     "bcryptjs": "^2.4.3",
+    "body-parser": "^1.18.3",
+    "cookie-parser": "^1.4.4",
     "cors": "^2.8.4",
     "cross-env": "^5.2.0",
     "dotenv": "^8.0.0",
