--- src/controllers/UserController.js
@@ -1,8 +1,0 @@
-/**
- * A collection of methods that controls user's interaction via the User routes
- *
- * @class UserController
- */
-class UserController {}
-
-export default UserController;
