+++ src/services/UserService.js
@@ -39,44 +39,13 @@ export default class UserService {
   }
 
   /**
-* @description Sign-in existing user
-* @param {object} userDetails user;s login details
-* @param {object} res response object
-* @returns {object}  JSON response
-*/
-  static async userLogin(userDetails, res) {
-    const { email, password } = userDetails;
-    if (!email || !password) {
-      return res.status(400).json({
-        status: 400,
-        error: 'Please provide complete login details'
-      });
-    }
-    try {
-      const user = await database.User.findOne({
-        where: { email }
-      });
-      if (!user) {
-        return res.status(401).json({
-          status: 401,
-          error: 'Invalid login details',
-        });
-      }
-      if (!Auth.compare(password, user.password)) {
-        return res.status(401).json({
-          status: 401,
-          error: 'Invalid password',
-        });
-      }
-      const { id, firstName, lastName } = user;
-      const output = { email, firstName, lastName };
-      const token = Auth.generateToken({ id, email });
-      return res.status(200).json({
-        status: 200,
-        data: { token, ...output }
-      });
-    } catch (error) {
-      return error;
-    }
+ * @description Sign-in existing user
+ * @param {object} email user's registered email
+ * @returns {Promise<object>} A promise object with user detail.
+ */
+  static userLogin(email) {
+    return database.User.findOne({
+      where: { email }
+    });
   }
-} 
+}
