+++ test/auth.spec.js
@@ -1,90 +1,116 @@
-import chai, { expect } from 'chai';
+import '@babel/polyfill';
+import chai from 'chai';
 import chaiHttp from 'chai-http';
 import faker from 'faker';
-import server from '../src';
-import { newUser } from './dummies';
-import Helpers from '../src/utils/helpers';
-
-const { generateToken } = Helpers;
+import app from '../src/index';
+import { helpers } from '../src/utils';
+import database from '../src/models';
+import Auth from '../src/utils/auth';
+import app from '../src/index';
 
 chai.use(chaiHttp);
-let newlyCreatedUser = {};
-describe('Auth route endpoints', () => {
-  it('should signup successfully with a status of 201', async () => {
-    const user = {
-      email: faker.internet.email(),
-      firstName: faker.name.firstName(),
-      lastName: faker.name.lastName(),
-      password: faker.internet.password(15, false),
-      companyName: faker.company.companyName(),
-      country: faker.address.country(),
-      gender: 'male',
-      street: faker.address.streetAddress(),
-      city: faker.address.city(),
-      state: faker.address.state(),
-      birthdate: faker.date.past(),
-      phoneNumber: faker.phone.phoneNumber()
-    };
+console.log(process.env.NODE_ENV);
+const { expect } = chai;
+let request;
+const user = {
+  firstName: 'Adebayo',
+  lastName: 'Daramola',
+  birthdate: '1986-09-21',
+  preferredLanguage: 'EN',
+  preferredCurrency: 'Naira',
+  email: 'ade.steve@gmail.com',
+  gender: 'Male',
+  street: 'Backstreet',
+  city: 'Ilupeju',
+  state: 'Lagos',
+  country: 'Nigeria',
+  zip: '100001',
+  phoneNo: '2347080445678',
+  companyName: 'Andela',
+  password: Auth.hash('testing'),
+  company: 'Andela',
+  role: 'Senior',
+  isVerified: false,
+  facebookId: 'Nil',
+  googleId: 'Nil',
+  department: 'admin',
+  lineManager: 'Lati',
+  createdAt: new Date(),
+  updatedAt: new Date()
+};
 
-    const response = await chai
-      .request(server)
-      .post('/api/auth/signup')
-      .send(user);
-    expect(response).to.have.status(201);
-    expect(response.body.data).to.be.a('object');
-    expect(response.body.data.token).to.be.a('string');
-    expect(response.body.data.firstName).to.be.a('string');
-    expect(response.body.data.lastName).to.be.a('string');
-  });
-  it("should send a verification link to a user's email upon successful registration", async () => {
-    const response = await chai.request(server).post('/api/auth/signup').send(newUser);
-    const { body: { data } } = response;
-    newlyCreatedUser = { ...data };
-    expect(response).to.have.status(201);
-    expect(data.isVerified).to.equal(false);
-    expect(data.emailSent).to.equal(true);
-  });
-
-  it('should return a response that indicates verification link was not sent whenever a mail is not sent upon successful registration', async () => {
-    const user2 = { ...newUser };
-    user2.email = 'notvalid';
-    const response = await chai.request(server).post('/api/auth/signup').send(user2);
-    const { body: { data } } = response;
-    newlyCreatedUser = { ...data };
-    expect(response).to.have.status(201);
-    expect(data.isVerified).to.equal(false);
-    expect(data.emailSent).to.equal(false);
-  });
+beforeEach(() => {
+  request = chai.request(app);
 });
-describe('GET /api/auth/verify?token', () => {
-  it('should successfully verify an existing user if the token is valid', async () => {
-    const { id, firstName, role } = newlyCreatedUser;
-    const token = generateToken({ id, firstName, role });
-    const response = await chai.request(server).get(`/api/auth/verify?token=${token}`);
-    const { body: { data: { isVerified } } } = response;
-    expect(response).to.have.status(200);
-    expect(isVerified).to.equal(true);
+
+describe('Auth route', () => {
+  before(async () => {
+    await database.sequelize.sync({ force: true });
   });
 
-  it('should return an error response if verification token is not provided', async () => {
-    const response = await chai.request(server).get('/api/auth/verify?token=');
-    const { body: { error } } = response;
-    expect(response).to.have.status(400);
-    expect(error.message).to.equal('Invalid token, verification unsuccessful');
+  after(async () => {
+    await database.sequelize.drop();
+    await database.sequelize.sync();
   });
 
-  it('should return an error response if verification token is not valid', async () => {
-    const response = await chai.request(server).get('/api/auth/verify?token=73783489d.eue4.78');
-    const { body: { error } } = response;
-    expect(response).to.have.status(400);
-    expect(error.message).to.equal('Invalid token, verification unsuccessful');
+  const user = {
+    email: faker.internet.email(),
+    firstName: faker.name.firstName(),
+    lastName: faker.name.lastName(),
+    password: helpers.hashPassword(faker.internet.password(15, false)),
+    companyName: faker.company.companyName(),
+    country: faker.address.country(),
+    gender: 'male',
+    street: faker.address.streetAddress(),
+    city: faker.address.city(),
+    state: faker.address.state(),
+    birthdate: faker.date.past(),
+    phoneNumber: faker.phone.phoneNumber()
+  };
+
+  describe('Signup route', () => {
+    it('should signup successfully with a status of 201', async () => {
+      const response = await request.post('/api/auth/signup').send(user);
+      expect(response.body.status).to.equal(201);
+      expect(response.body.data).to.be.a('object');
+      expect(response.body.data.token).to.be.a('string');
+      expect(response.body.data.firstName).to.be.a('string');
+      expect(response.body.data.lastName).to.be.a('string');
+    });
   });
 
-  it("should return an error response if the verification link contains details of a user that doesn't exist", async () => {
-    const token = generateToken({ id: 49 });
-    const response = await chai.request(server).get(`/api/auth/verify?token=${token}`);
-    const { body: { error } } = response;
-    expect(response).to.have.status(400);
-    expect(error.message).to.equal('no user found to verify');
+  describe('Login route', () => {
+    it('should signin successfully with a status of 200', async () => {
+      const login = {
+        email: user.email,
+        password: user.password,
+      };
+      const response = await request.post('/api/auth/login').send(login);
+      expect(response.body.status).to.equal(200);
+      expect(response.body.data).to.be.a('object');
+      expect(response.body.data.token).to.be.a('string');
+      expect(response.body.data.firstName).to.be.a('string');
+      expect(response.body.data.lastName).to.be.a('string');
+    });
+
+    it('should return 401 if password is invalid', async () => {
+      const login = {
+        email: user.email,
+        password: 'wrong',
+      };
+      const response = await request.post('/api/auth/login').send(login);
+      expect(response.body.status).to.equal(401);
+      expect(response.body.message).to.be.equal('Invalid login details');
+    });
+
+    it('should return 401 error if login details is invalid', async () => {
+      const login = {
+        email: 'kylewalker123@yahoo.com',
+        password: 'password'
+      };
+      const response = await request.post('/api/auth/login').send(login);
+      expect(response.body.status).to.equal(401);
+      expect(response.body.message).to.be.equal('Invalid login details');
+    });
   });
 });
